Creating Player in Home
Player made static and player calls in packet parser made to reference
the static player to allow creation of networking components to be
completed in the create/join room methods.
Also attempted saving player name, failed as code required needs to be
in a fragment and unsure how they work atm.